# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shoken WebApp API",
    "description": "証券管理Webアプリ APIドキュメント",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/stocks": {
      "get": {
        "tags": [
          "handlers::v1::stocks"
        ],
        "summary": "銘柄情報を検索（v1: クエリパラメータ）",
        "operationId": "v1_stock_search",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "銘柄コードまたは銘柄名",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stock"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "handlers::v1::stocks"
        ],
        "summary": "銘柄情報を追加（v1）",
        "operationId": "v1_stock_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Stock"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stock"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/stocks": {
      "post": {
        "tags": [
          "handlers::stock"
        ],
        "summary": "銘柄情報を追加（認証必須）",
        "operationId": "stock_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Stock"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stock"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetails": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        }
      },
...
      "Stock": {
        "type": "object",
        "required": [
          "date",
          "code",
          "name",
          "market_category"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "industry_category_17": {
            "type": [
              "string",
              "null"
            ]
          },
          "industry_category_33": {
            "type": [
              "string",
              "null"
            ]
          },
          "industry_code_17": {
            "type": [
              "string",
              "null"
            ]
          },
          "industry_code_33": {
            "type": [
              "string",
              "null"
            ]
          },
          "market_category": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "size_category": {
            "type": [
              "string",
              "null"
            ]
          },
          "size_code": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token"
      }
    }
  }
}